RNAProduct: Silence perlcritic about @_ unpacking in get_all_DBEntries() aliases
In this particular case we *do* very much want to pass the argument
array to the proper method as it is, both in order to minimise the
overhead of aliases and just in case someone decides in the end to e.g.
to take advantage of direct modification of @_ contents in
get_all_DBEntries().