stable_id_lookup: extract RNAProduct stable IDs from core databases
Uses the same type of SQL SELECT queries as Translation, which makes
sense given how similar they are.

Tested on test-genome-DBs/homo_sapiens/core, works without errors.

Aborts upon encountering a core database missing the 'rnaproduct' table
but that is in my humble opinion very much desired behaviour, as it could
indicate incomplete application of schema patches in the release this
will be included in.