MicroRNA: having multiple mirna_arms in the DB is now fatal
On the one hand there seems to be no good reason for a MicroRNA object
having multiple arm attributes, on the other hand simply taking the
first one in the event of it happening may not even return the same
value every time (it depends on a lot of factors, from the
implementation details of AttributeAdaptor up to and including the
underlying database system). There was a warning but who reads those
anyway. Just throw an exception instead, if someone decides to allow
such a case they can always catch it.